Checked the .xml documents, went through the tests, etc.
All schedules should be good now. There was a slightly weird thing where a test was failing with the cybersecurity cert, but that was the only small thing I found. I double-checked the actual schedule and the schedule was right, so I probably just messed up that particular test.

Either way: THE OR FUNCTION WORKS, a random class is taken.